refactor(oidc): separate metadata-backed clients
Keep RFC 7591 registration and dcr_ encoding in dcr.go. Move the normalized DCR/CIMD metadata model, shared redirect validation and Zitadel client implementation into metadata_client.go so each protocol file has a clear boundary.
